refactor($result): 게임의 결과를 나타내는 모든 클래스를 패키지로 묶는다
- ScoreBoard, PlayerWinning, DealerWinning, WinningResultStatus
