+++ src/main/kotlin/blackjack/model/result/DealerWinning.kt
@@ -1,3 +1,3 @@
-package blackjack.model
+package blackjack.model.result
 
 data class DealerWinning(val result: Map<WinningResultStatus, Int>)
