+++ src/main/kotlin/blackjack/model/result/ScoreBoard.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.result
 
 import blackjack.model.role.PlayerName
