+++ src/test/kotlin/blackjack/model/WinningResultStatusTest.kt
@@ -1,5 +1,6 @@
 package blackjack.model
 
+import blackjack.model.result.WinningResultStatus
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
